# Conflicts:
#	README.md
#	cmake/ospray_version.cmake
#	modules/mpi/MPIOffloadDevice.cpp

+++ ospray/api/ISPCDevice.cpp
@@ -216,7 +216,7 @@ namespace ospray {
                              const bool b)
     {
       ManagedObject *object = (ManagedObject *)_object;
-      object->setParam(bufName, b);
+      object->setParam(bufName, static_cast<int>(b));
     }
 
     void ISPCDevice::setFloat(OSPObject _object,
